# specs/002-rag-chatbot/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: RAG Chatbot API
  version: 1.0.0
paths:
  /health:
    get:
      summary: Health Check
      responses:
        '200':
          description: OK
  /query:
    post:
      summary: Ask a question to the chatbot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                conversation_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: The chatbot's answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                  sources:
                    type: array
                    items:
                      type: string
                  conversation_id:
                    type: string
                    format: uuid
  /selected-text:
    post:
      summary: Ask a question about a selected piece of text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                text:
                  type: string
                conversation_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: The chatbot's answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                  sources:
                    type: array
                    items:
                      type: string
                  conversation_id:
                    type: string
                    format: uuid
  /ingest:
    post:
      summary: Trigger the ingestion of documents
      responses:
        '202':
          description: Ingestion job started
  /conversation:
    get:
      summary: Get conversation history
      parameters:
        - name: conversation_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A list of messages in the conversation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    role:
                      type: string
                    content:
                      type: string
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []